requirements.txt
torch
torchvision
python-box==6.0.2
numpy
pandas
matplotlib
seaborn
tqdm
pyyaml
ensure
joblib
scipy
fastapi
pydantic
scipy
kaggle
kagglehub
python-dotenv
loguru
ipython
opencv-python
-e .